+++ typecats/__init__.py
@@ -1,1 +1,5 @@
 from .tc import Cat, unstruc, struc, register_struc_hook, register_unstruc_hook  # noqa
+
+from .__about__ import __version__
+
+__all__ = [__version__]